+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix issue - Joined geometry does not convert correctly
+
 ## 0.1.13
 * update DynamoCore to 2.5.0.7186